sbi: error and function get their own module
To allow further changes to break up sbi.rs into components that will
each depend on error and function.
